Commit
This commit introduces "capabilities" to comlink. You can now forward received proxies to other locations and new message channels are set up in the background to handle the comms. Technically I've introduced this in an earlier commit, but this one makes it so that the proxy that was forwarded doesn't get disconnected in the correct context by default. It uses the pre-existing `createEndpoint` routine to accomplish this, but I've fixed it, so that it returns a new message port immediately. Effectively, if you forward a proxy to a 3rd location, message ports are sent in both directions. Since an object can now be exposed to multiple locations, we also need to maintain a counter and only run the object's dispose/release method once that number reaches 0. Technically, this was this could happen before, but now it is fixed.